+++ src/state.py
@@ -3,7 +3,6 @@
 
 """Jenkins States."""
 import dataclasses
-import functools
 import logging
 import os
 import typing
@@ -121,27 +120,13 @@ class AgentMeta(BaseModel):
         return cls(executors=num_executors, labels=labels, name=name)
 
 
-def _is_remote_unit(app_name: str, unit: ops.Unit) -> bool:
-    """Return whether the unit is a remote unit in a relation.
-
-    Args:
-        app_name: The current application name.
-        unit: The unit to check in a relation.
-
-    Returns:
-        True if is remote unit. False otherwise.
-    """
-    return app_name not in unit.name
-
-
 def _get_agent_meta_map_from_relation(
-    relations: typing.List[ops.Relation], current_app_name: str
+    relations: typing.List[ops.Relation],
 ) -> typing.Optional[typing.Mapping[ops.Relation, list[AgentMeta]]]:
     """Return a mapping of unit name to AgentMetadata from agent relation.
 
     Args:
         relations: The agent relations.
-        current_app_name: Current application name, i.e. "jenkins-k8s-operator".
 
     Returns:
         A mapping of ops.Unit to AgentMetadata.
@@ -150,8 +135,7 @@ def _get_agent_meta_map_from_relation(
         return None
     relation_agents_map = {}
     for relation in relations:
-        remote_units = filter(functools.partial(_is_remote_unit, current_app_name), relation.units)
-        agents = [AgentMeta.from_agent_relation(relation.data[unit]) for unit in remote_units]
+        agents = [AgentMeta.from_agent_relation(relation.data[unit]) for unit in relation.units]
         relation_agents_map.update({relation: [agent for agent in agents if agent]})
     return relation_agents_map
 
@@ -246,7 +230,7 @@ class State:
 
         try:
             agent_relation_meta_map = _get_agent_meta_map_from_relation(
-                charm.model.relations[AGENT_RELATION], charm.app.name
+                charm.model.relations[AGENT_RELATION]
             )
             is_auth_proxy_integrated = _is_auth_proxy_integrated(
                 charm.model.get_relation(AUTH_PROXY_RELATION)